feat: move get-formatted-countdown-units to separate helper bb-321
